chore(ui/roles): move role names to a separate file
