+++ ui/src/admin/components/chronograf/OrganizationsTableRowNew.js
@@ -5,7 +5,7 @@ import ConfirmOrCancel from 'shared/components/ConfirmOrCancel'
 import Dropdown from 'shared/components/Dropdown'
 
 import {USER_ROLES} from 'src/admin/constants/chronografAdmin'
-import {MEMBER_ROLE} from 'src/auth/Authorized'
+import {MEMBER_ROLE} from 'src/auth/roles'
 import {ErrorHandling} from 'src/shared/decorators/errors'
 
 class OrganizationsTableRowNew extends Component {
